requirements_final.txt
# Facial Recognition System - Apple Silicon
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.8.1.78
Flask==2.3.3
tensorflow-macos==2.13.0
tensorflow-metal==1.0.1
keras==2.13.1
deepface==0.0.75
protobuf==3.20.3
pandas>=1.3.0
gdown>=4.6.0
tqdm>=4.64.0
